Created a ContextValidatorFactory and moved the createNonStrictRegistrationContextValidator method to this class: create single entry point for context validator creation and - maybe - implement a single instance of the Json and Cbor mappers.
