+++ webauthn4j-core/src/main/java/com/webauthn4j/validator/WebAuthnContextValidatorFactory.java
@@ -1,0 +1,69 @@
+package com.webauthn4j.validator;
+
+import com.webauthn4j.converter.util.CborConverter;
+import com.webauthn4j.converter.util.JsonConverter;
+import com.webauthn4j.validator.attestation.statement.androidkey.AndroidKeyAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.androidkey.NullAndroidKeyAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.androidsafetynet.AndroidSafetyNetAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.androidsafetynet.NullAndroidSafetyNetAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.none.NoneAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.packed.NullPackedAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.packed.PackedAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.tpm.TPMAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.u2f.FIDOU2FAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.statement.u2f.NullFIDOU2FAttestationStatementValidator;
+import com.webauthn4j.validator.attestation.trustworthiness.certpath.NullCertPathTrustworthinessValidator;
+import com.webauthn4j.validator.attestation.trustworthiness.ecdaa.NullECDAATrustworthinessValidator;
+import com.webauthn4j.validator.attestation.trustworthiness.self.NullSelfAttestationTrustworthinessValidator;
+
+import java.util.Arrays;
+
+public final class WebAuthnContextValidatorFactory {
+    private WebAuthnContextValidatorFactory() {
+
+    }
+
+    public static WebAuthnRegistrationContextValidator createNonStrictRegistrationContextValidator() {
+        return createNonStrictRegistrationContextValidator(new JsonConverter(), new CborConverter());
+    }
+
+    public static WebAuthnRegistrationContextValidator createNonStrictRegistrationContextValidator(JsonConverter jsonConverter, CborConverter cborConverter) {
+        return new WebAuthnRegistrationContextValidator(
+                Arrays.asList(
+                        new NoneAttestationStatementValidator(),
+                        new NullFIDOU2FAttestationStatementValidator(),
+                        new NullPackedAttestationStatementValidator(),
+                        new NullAndroidKeyAttestationStatementValidator(),
+                        new NullAndroidSafetyNetAttestationStatementValidator()
+                ),
+                new NullCertPathTrustworthinessValidator(),
+                new NullECDAATrustworthinessValidator(),
+                new NullSelfAttestationTrustworthinessValidator(),
+                jsonConverter,
+                cborConverter
+        );
+    }
+
+    public static WebAuthnRegistrationContextValidator createDefaultRegistrationContextValidator(JsonConverter jsonConverter, CborConverter cborConverter) {
+        return new WebAuthnRegistrationContextValidator(
+                Arrays.asList(
+                        new NoneAttestationStatementValidator(),
+                        new FIDOU2FAttestationStatementValidator(),
+                        new PackedAttestationStatementValidator(),
+                        new AndroidKeyAttestationStatementValidator(),
+                        new AndroidSafetyNetAttestationStatementValidator(),
+                        new TPMAttestationStatementValidator()
+                ),
+                new NullCertPathTrustworthinessValidator(),
+                new NullECDAATrustworthinessValidator(),
+                new NullSelfAttestationTrustworthinessValidator(),
+                jsonConverter,
+                cborConverter
+        );
+    }
+
+    public static WebAuthnAuthenticationContextValidator createDefaultAuthenticationContextValidator() {
+        return new WebAuthnAuthenticationContextValidator(new JsonConverter(), new CborConverter());
+    }
+
+}
